fix(briefings): additional code review follow-ups for phase 2
- Extract BriefingOptions to shared/types/briefing-options
- Replace console.log/error with NestJS Logger in generation service
- Add limit query parameter to list briefings endpoint
