+++ src/shared/types/briefing-options.ts
@@ -1,0 +1,11 @@
+import { FeedProfile } from './feed';
+
+export interface BriefingOptions {
+  feedProfile?: FeedProfile;
+  lookbackHours?: number;
+  minArticles?: number;
+  customPrompts?: Partial<{
+    clusterAnalysis: string;
+    briefSynthesis: string;
+  }>;
+}
